requirements.txt
git+https://github.com/django-extensions/django-extensions.git@25a41d8a3ecb24c009c5f4cac6010a091a3c91c8

asgiref==3.5.2
astroid==2.5
atomicwrites==1.4.1
attrs==22.1.0
black==22.3.0
cachetools==5.2.0
certifi==2022.6.15
cfgv==3.3.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
cssselect==1.1.0
cssutils==2.6.0
dislib==0.4.3
distlib==0.3.5
Django==3.2.9
django-admin-generator==2.2.1
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.14.0
django-cors-middleware==1.5.0
django-developmentEmailDashboard==2.0.2
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-utils2==2.12.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dlib==19.22.0
emails==0.6
face-recognition==1.3.0
filelock==3.8.0
gunicorn==20.1.0
identify==2.5.1
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
lazy-object-proxy==1.7.1
lxml==4.9.1
Markdown==3.3.4
MarkupSafe==2.1.1
martor==1.6.13
mccabe==0.6.1
mypy==0.971
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.23.2
openapi-codec==1.3.2
opencv-python
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.19.0
premailer==3.10.0
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.4.0
pylint==2.7.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-utils==3.3.3
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.0
simplejson==3.17.6
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.14.1
werkzeug==2.2.2
wrapt==1.12.1